% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{as.matrix.mcgraph}
\alias{as.matrix.mcgraph}
\title{Extract the adjacency matrix from a mcgraph graph object.}
\usage{
\method{as.matrix}{mcgraph}(x, ...)
}
\arguments{
\item{x}{A mcgraph object.}

\item{\dots}{currently not used.}
}
\value{
The adjacency matrix without other mcgraph attributes.
}
\description{
Extract the adjacency matrix from a mcgraph graph object and remove all other attributes.
}
\examples{
band=mcg.band(nodes=12)
class(band)
adj=as.matrix(band)
is.mcgraph(adj)
is.matrix(adj)
class(adj)
}
\author{
Detlef Groth <email: dgroth@uni-potsdam.de>
}
