% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{degree}
\alias{degree}
\title{Degree of `mcgraph` graphs}
\usage{
degree(x, ...)
}
\arguments{
\item{x}{Object of class igraph, mcgraph or something else}

\item{\dots}{Arguments forwarded to the right namespace method.}
}
\description{
Get the number of edges for vertices of a graph. Generic function.plots mcgraph objects
}
\seealso{
\link[mcgraph:degree.mcgraph]{degree.mcgraph}
}
