% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{density.mcgraph}
\alias{density.mcgraph}
\title{Graph density of `mcgraph` objects}
\usage{
\method{density}{mcgraph}(x, ...)
}
\arguments{
\item{x}{object of class mcgraph.}

\item{\dots}{parameters which are forwarded to the generic density function. Currently not used.}
}
\description{
`density.mcgraph` gives back the graph density of the graph object.
}
\details{
The graph density of a graph is the ratio of the number of edges divided by the  number of possible edges.
}
\examples{
bara=mcg.barabasi(nodes=16,m=2)
density(bara)
}
\author{
: Detlef Groth <dgroth@uni-potsdam.de>
}
