% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{is.mcgraph}
\alias{is.mcgraph}
\title{Type check for `mcgraph` objects}
\usage{
is.mcgraph(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical value. TRUE if argument is a mcgraph object.
}
\description{
Checks, whether an object is of type `mcgraph`.
}
\examples{
band=mcg.band(nodes=12)
is.mcgraph(band)
is.mcgraph(1)
}
\author{
Detlef Groth <email: dgroth@uni-potsdam.de>
}
