% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.band}
\alias{mcg.band}
\title{Creation of band or chain graph}
\usage{
mcg.band(nodes = 8)
}
\arguments{
\item{nodes}{number of nodes to create.}
}
\value{
graph of class mcgraph with adjacency matrix for an undirected graph
}
\description{
`mcg.band` creates a band, alias chain graph with given number of nodes
}
\details{
Create a simple band or chain graph with a given number of nodes.
}
\examples{
band=mcg.band(12)
plot(band,vertex.color="salmon")
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>
}
