% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.circular}
\alias{mcg.circular}
\title{Creation of circular graphs}
\usage{
mcg.circular(nodes = 8)
}
\arguments{
\item{nodes}{number of nodes to create in the graph}
}
\value{
graph of class mcgraph with adjacency matrix for an undirected graph
}
\description{
`mcg.circular` creates a circular graph with given number of nodes
}
\details{
This function create a simple circular graph with a given number of nodes. All nodes in the graph have degree two.
}
\examples{
circ=mcg.circular(12)
plot(circ,layout='circle')
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>
}
