% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.cluster}
\alias{mcg.cluster}
\title{Creation of densely connected graph components}
\usage{
mcg.cluster(nodes = 26, cluster = 2, edges = 50)
}
\arguments{
\item{nodes}{Number of nodes in the graph. Default: 26}

\item{cluster}{Number of clusters, graph components. Default: 2}

\item{edges}{Number of edges in the graph.}
}
\value{
Graph of class mcgraph with adjacency matrix for an undirected graph.
}
\description{
`mcg.cluster` creates several graph components which are highly connected within each component.
}
\details{
This function create a random graph whith a given number of nodes and a given cluster number.
Nodes are only connected to other nodes of the same cluster,
}
\examples{
clust=mcg.cluster(nodes=30,cluster=3)
plot(clust,vertex.color=rainbow(20))
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>
}
