% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.cross}
\alias{mcg.cross}
\title{Creation of graphs with a central node and outgoing bands}
\usage{
mcg.cross(bands = 4, length = 4)
}
\arguments{
\item{bands}{number of bands outgoing from the central node}

\item{length}{length of nodes number of nodes to create}
}
\value{
graph of class mcgraph with adjacency matrix for an undirected graph
}
\description{
`mcg.cross` creates a central node with several outgoind bands of same
length.
}
\details{
This graph generation function will generated graphs with one centreal node and several chains of same length attached to the central node.
The final number of nodes will be bands x length + 1.
}
\examples{
cross=mcg.cross(bands=8,length=5)
plot(cross)
}
