% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.d2u}
\alias{mcg.d2u}
\title{Conversion of directed to undirected graphs}
\usage{
mcg.d2u(A)
}
\arguments{
\item{A}{directed input graph}
}
\value{
undirected mcgraph object
}
\description{
`mcg.d2u` gets a directed input graph and returns an undirected graph where every edges becomes undirected.
}
\examples{
cross=mcg.cross(bands=5,length=5)
D=mcg.u2d(cross,input=3)
U=mcg.d2u(D)
par(mfrow=c(1,2))
plot(D)
plot(U)
}
\author{
: Detlef Groth <dgroth@uni-potsdam.de>
}
