% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.random}
\alias{mcg.random}
\title{Creation of random graphs}
\usage{
mcg.random(nodes = 26, edges = 50)
}
\arguments{
\item{nodes}{Number of nodes in the graph.}

\item{edges}{Number of edges in the graph.}
}
\value{
Graph of class mcgraph with adjacency matrix for an undirected graph.
}
\description{
`mcg.random` creates a random graph for given number of nodes and edges
}
\details{
This function create a random graph whith a given number of nodes and edges.
All possible node pairs the same probability of being connected by an edge.
}
\examples{
rand=mcg.random(nodes=20,edges=30)
plot(rand,vertex.color=rainbow(20))
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>
}
