% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{mcg.special}
\alias{mcg.special}
\title{Creation of special graphs}
\usage{
mcg.special(gname)
}
\arguments{
\item{gname}{Name of the graph to get. Possible names are currently: hub40, hub04, hub31, hub13, hub22,
hub30, hub03, hub21, hub12, nicolas, werner, wernerr, wernerc.}
}
\value{
directed graph of class mcgraph.
}
\description{
`mcg.special` creates some very special directed graphs.
}
\details{
This function just returns some interesting small graphs for playing and analysing.
}
\examples{
g40=mcg.special("hub40");
plot(g40,vertex.color=rainbow(5))
nic=mcg.special("nicolas")
plot(nic,vertex.color=c("skyblue",rep("salmon", 4)),vertex.size=4,
     edge.width=3,arr.length=0.2,label.size=2,
     edge.color="black")
}
\author{
Detlef Groth <dgroth@uni-potsdam.de>
}
