% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{mcgraph-package}
\alias{mcgraph}
\alias{mcgraph-package}
\title{mcgraph: Data Generation for Graph Reconstruction Using Monte Carlo Sampling}
\description{
Create various graph types and generate synthetic data using a simple Monte Carlo sampling approach. Sampling can be based on defined input graphs, either directed, undirected, with and without weights or on given data. Additionally, functions for graph reconstruction, assessment of graph reconstructs and graph visualization are provided. At the moment, we mainly concentrate on small to medium sized graphs up until around approximately 1000 nodes for which the runtimes of the functions should be in a reasonable range.
}
\details{
The only function you're likely to need from cat is [cat()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Detlef Groth \email{dgroth@uni-potsdam.de} (\href{https://orcid.org/0000-0002-3441-3978}{ORCID})

Authors:
\itemize{
  \item Masiar Novine \email{novinefarahb@uni-potsdam.de} (\href{https://orcid.org/0000-0001-9687-8675}{ORCID})
}

}
\keyword{internal}
