% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgraph.R
\name{summary.mcgraph}
\alias{summary.mcgraph}
\title{summary.mcgraph gives back summary of mcgraph properties}
\usage{
\method{summary}{mcgraph}(object, ...)
}
\arguments{
\item{object}{object of class mcgraph.}

\item{\dots}{parameters which are forwarded to the generic summary functions. Currently not used.}
}
\description{
`summary.mcgraph` gives back summary of mcgraph properties.
}
\details{
This function gives back the graph properties such us type of graph, number of edges and nodes.
}
\examples{
band=mcg.band(16)
summary(band)
}
\author{
: Detlef Groth <dgroth@uni-potsdam.de>
}
